pyproject.toml
[project]
name = "Project8-API"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.29.0",
    "sqlalchemy==2.0.30",
    "alembic==1.13.1",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "hypercorn (>=0.17.3,<0.18.0)",
    "requests (>=2.32.5,<3.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
pytest-cov = "^5.0.0"
python-dotenv = "^1.0.1"
pre-commit = "^4.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
